modified xsd schema version

+++ org.codegen.libs/trunk/org.codegen.lib/Model/IAuditable.vb
@@ -23,7 +23,7 @@ Public Interface IAuditable2
     Property CreateDate() As DateTime?
     Property UpdateDate() As DateTime?
 
-    Property CreateUser() As Integer?
-    Property UpdateUser() As Integer?
+    Property CreateUser() As Long?
+    Property UpdateUser() As Long?
 
 End Interface
